moved session utils file
