+++ src/app/core/session/session-states/session-utils.ts
